@@ -1,23 +1,21 @@
 {
+  "$schema": "https://json.schemastore.org/package",
   "name": "vue-router-mock",
-  "packageManager": "pnpm@10.6.5",
+  "packageManager": "pnpm@10.14.0",
   "version": "1.1.0",
   "description": "Easily test your components by mocking the router",
-  "main": "dist/index.cjs",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "exports": {
-    "types": "./dist/index.d.ts",
-    "require": "./dist/index.cjs",
-    "import": "./dist/index.mjs"
-  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "type": "module",
+  "types": "./dist/index.d.ts",
+  "exports": "./dist/index.js",
   "sideEffects": false,
   "author": {
     "name": "Eduardo San Martin Morote",
     "email": "posva13@gmail.com"
   },
   "scripts": {
-    "build": "unbuild",
+    "build": "tsdown",
     "release": "publie",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 1",
     "lint": "prettier -c --parser typescript \"{src,__tests__,e2e}/**/*.[jt]s?(x)\"",
@@ -28,9 +26,7 @@
     "prepare": "simple-git-hooks"
   },
   "files": [
-    "dist/*.cjs",
-    "dist/*.mjs",
-    "dist/*.d.ts",
+    "./dist",
     "LICENSE",
     "README.md"
   ],
@@ -47,23 +43,23 @@
   "license": "MIT",
   "devDependencies": {
     "@types/jest": "^30.0.0",
-    "@types/sinon": "^17.0.3",
-    "@vitest/coverage-v8": "^3.0.2",
+    "@types/sinon": "^17.0.4",
+    "@vitest/coverage-v8": "^3.2.4",
     "@vue/test-utils": "^2.4.6",
-    "chalk": "^5.3.0",
+    "chalk": "^5.5.0",
     "conventional-changelog-cli": "^5.0.0",
     "enquirer": "^2.4.1",
-    "execa": "^9.3.1",
+    "execa": "^9.6.0",
     "happy-dom": "^18.0.1",
-    "lint-staged": "^16.0.0",
+    "lint-staged": "^16.1.4",
     "minimist": "^1.2.8",
-    "prettier": "^3.3.3",
+    "prettier": "^3.6.2",
     "publie": "^0.0.2",
-    "semver": "^7.6.3",
-    "simple-git-hooks": "^2.11.1",
-    "typescript": "^5.5.4",
-    "unbuild": "^3.0.1",
-    "vitest": "^3.0.2",
+    "semver": "^7.7.2",
+    "simple-git-hooks": "^2.13.1",
+    "tsdown": "^0.13.2",
+    "typescript": "^5.9.2",
+    "vitest": "^3.2.4",
     "vue": "^3.4.38",
     "vue-router": "^4.4.3"
   },
